elevenlabs-t2s-mcp/pyproject.toml
[project]
name = "elevenlabs-t2s-mcp"
version = "0.1.0"
description = "MCP server for ElevenLabs text-to-speech"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
    "elevenlabs>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[project.scripts]
elevenlabs-t2s = "elevenlabs_t2s_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/elevenlabs_t2s_mcp"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
per-file-ignores = {"src/elevenlabs_t2s_mcp/server.py" = ["E501"]}

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
